+++ ReactNativeClient/package.json
@@ -54,7 +54,7 @@
     "react-async": "^10.0.0",
     "react-native": "0.61.5",
     "react-native-action-button": "^2.6.9",
-    "react-native-audio-recorder-player": "^2.5.0",
+    "react-native-audio-recorder-player": "^2.5.1",
     "react-native-camera": "^2.10.2",
     "react-native-datepicker": "^1.6.0",
     "react-native-device-info": "^5.5.1",
@@ -96,6 +96,7 @@
   "devDependencies": {
     "@babel/core": "^7.6.2",
     "@babel/runtime": "^7.6.2",
+    "@types/react-native": "^0.62.10",
     "app-module-path": "^2.2.0",
     "execa": "^4.0.0",
     "fs-extra": "^8.1.0",
